refactor: :recycle: project components moved out of pages folder
