+++ components/Projects/RegularProject.tsx
